Throw more descriptive errors when getting a token
When Google responded with an error in response to a
token request, it was only throwing the status line
of the HTTP response.  We also want the actual error
supplied in the content of the response.

(cherry picked from commit b923b2e3e144ec3f6fa975358c747e269a4ef70b)
Signed-off-by: Nicolas R <nicolas@atoomic.org>